Fix value substitution and graph pruning
Substituting a value for another would result in the
new value being marked as "prunable", leading to an
inconsistent graph after pruning. Reorganize substitution
and pruning to fix that, at the same time use an iterative
pruning algorithm instead of a recursive one.

Also, add a test for the previously existing bug.