Remove cuts in region if undoing placement of anchor node
Nodes in sub-ordinate regions may only be placed when
the anchor node of the region has been placed as well.
Undoing the placement of the anchor node (e.g. when
destroying the node) must therefore cause the placement
of the nodes in the anchored regions to be undone as
well.