Specialize register classes before and during coloring
Locate two-operand instructions (that overwrite one of
their input operands with the output value) where either
input or output are a specialized register class allowing
only a single register. Specialize the matching output
or input to guide the register allocator in choosing
the same register. This considerably reduces later
two-operand fixup, usually to the point that no further
fixup is necessary.