regalloc: make auxnodes copyable
Implement "create" operation for auxnodes to make them copyable.
Actually there is no point in ever copying a (sub-)graph
containing auxnodes, as they are only used during shaping of the
graph and there is no way to copy a "shape-in-progress", but the
method should be implemented for completeness' sake.