regalloc: fetch cross-count before deciding on register
Fetch cross-count for the variable in question before deciding on
a specific register. This fixes a case where choice of a particular
register class at one point in the graph when an interfering
variable is restricted to a specific class but has not been
assigned a register yet (so technically, the name is not denied
to the candidate register).