regalloc: fix use-after-free in ssavar iteration
When merging ssavars from one subregion into its parents, the list
of ssavars active at the top of the inner region is being
iterated. Depending however on the merge strategy picked per
ssavar, the crossing-point with the inner region might cease to
exist after the merge. Consequently we are iterating a list where
the current element might end up deleted and must therefore
pick the _SAFE variant of the iterators to avoid a potential
use-after-free.