vsdg: change reduction interface for operators
Change interface for performing reductions on unary and binary
operators: have "can_reduce" return a "hint" as to what kind of
reduction can successfully be performed, while the reduction
itself can use this hint to avoid checking the same preconditions.
As a result, calling "can_reduce" before "reduce" is now mandatory,
the methods are orthogonal and "tentatively" calling "reduce"
is not supported any longer.

Signed-off-by: Helge Bahmann <hcb@chaoticmind.net>