vsdg: reverse semantic of branches to gamma node
Up until now, the semantic of the two alternative branches to the
gamma node has been that the "left" (or "first") branch corresponds
to the "false" evaluation of the control predicate, while the
"second" (or "right") one corresponds to the "true" evaluation.
This has turned out very confusing since practically everyone
intuititively expects it precisely the other way around. Furthermore,
the way branches are fixed up now makes the whole reason for
this initial design decision moot.