arch: remove save_state and reset compilate functions
Remove the "save_state" and "reset" pair of operations that were previously
used to preserve the contents of the compilation target buffers. This model is
going to be a bit difficult for the new structure, so remove this functionality
for now and document that the buffer contents are destroyed by the
"compile" function.