vsdg: remove address field from labels
Remove the address field from labels, as labels are more abstract now and the
addresses are provided through the linker_symbol abstraction instead. Remove
the address field from the constructor, as well as all the fake addresses that
were made up at the call sites to satisfy the previous interface.

Signed-off-by: Nico Reißmann <nico.reissmann@gmail.com>