arch: linker symbols
Introduce abstract concept of a "linker symbol", as well as references to
symbols and section starts. Define interface to resolve the addresses of linker
symbols. Change relocation to be based on the concept of linker symbols and
references to these, and use resolver interface to provide the addresses for
the linking process.

Signed-off-by: Nico Reißmann <nico.reissmann@gmail.com>