vsdg: add methods to jive::operation, mirroring jive_node_class
Add methods to jive::operation and the more specialized unary and binary
operators that makes them self-describing: This includes declaring their
type signature, comparison, node instantiation and generating a human-readable
description.

Alter accessors such that a nullptr entry in the function pointer tables
of jive_node_class and unary/binary operator classes indicates that the
corresponding overridden function of jive::operation (or its derivative)
is to be used in its place.

Signed-off-by: Nico Reißmann <nico.reissmann@gmail.com>