tests: report errors on memory leak in instrumented tests
Run valgrind with --leak-check=full in the valgrind-test target. This appears
to be the only way to induce valgrind into returning a non-zero exit code
on detecting a memory leak.

The majority of tests work fine, there are a two known errors caused by the
incomplete transition to exceptions and exception-safe formulation of
memory management.

Signed-off-by: Nico Reißmann <nico.reissmann@gmail.com>