vsdg: move resource class requirements to generic code
Move support for specifying resource class requirements for inputs and outputs
to the generic operator interface. Let generic code query this interface and
set up requirements. This is currently required due to the tight coupling of
register allocation with generic code and will be removed through disentangling
of the register allocator later.

Signed-off-by: Nico Reißmann <nico.reissmann@gmail.com>