vsdg: gate inputs of graph tail node
The export function just added bluntly inputs to the tail node. We
cannot do this any longer if we want to remove inputs and outputs,
since the graph tail node has officially no arguments. This entire
mess is going to be solved once we got rid off head and tail nodes.

Signed-off-by: Helge Bahmann <hcb@chaoticmind.net>